feat: Delete ACK'd messages from storage to prevent some dupes
Closes: SYNC-4466
